# add faster reshape utility function