The old implementation was kind of weird, and relied on creating a dummy onp array (with stride tricks so memory wasn't wasted).

Here's the old way:

```python
def _compute_newshape(a, newshape):
  dummy = onp.broadcast_to(0, a.shape)
  return onp.reshape(dummy, newshape).shape
```

```
In [1]: import numpy as np

In [2]: x = np.ones((2, 2))

In [3]: %timeit _compute_newshape(x, (2, 1, 2))
4.77 µs ± 203 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [4]: %timeit _compute_newshape(x, (2, -1, 2))
4.52 µs ± 14.4 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

Here's the new way:

```python
def _compute_newshape(a, newshape):
  newsize = prod(newshape)
  if newsize < 0:
    fix = a.size // -newsize
    return [d if d != -1 else fix for d in newshape]
  else:
    return newshape
```

```
In [1]: import numpy as np

In [2]: x = np.ones((2, 2))

In [3]: %timeit _compute_newshape(x, (2, 1, 2))
218 ns ± 0.866 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %timeit _compute_newshape(x, (2, -1, 2))
467 ns ± 2.45 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```